+++ lib/rubocop/cop/style/percent_literal_delimiters.rb
@@ -59,11 +59,10 @@ module RuboCop
           end
         end
 
-        def on_percent_literal(node, types)
+        def on_percent_literal(node)
           type = type(node)
-          return unless types.include?(type) &&
-            !uses_preferred_delimiter?(node, type) &&
-            !contains_preferred_delimiter?(node, type)
+          return if uses_preferred_delimiter?(node, type) ||
+            contains_preferred_delimiter?(node, type)
 
           add_offense(node, :expression)
         end